backend/requirements.txt
Flask
Flask-RESTful
Flask-Mail
PyMongo
python-dotenv
langchain-mistralai  # For integrating Mistral with LangChain
requests
